+++ tests/test_multi_threaded_cleanup_use_utc.rs
@@ -68,9 +68,9 @@ mod d {
         cond_sync: &CondSync<usize>,
     ) -> Vec<JoinHandle<u8>> {
         let mut worker_handles: Vec<JoinHandle<u8>> = Vec::with_capacity(no_of_workers);
-        trace!("Starting {} worker threads", no_of_workers);
+        trace!("Starting {no_of_workers} worker threads");
         for thread_number in 0..no_of_workers {
-            trace!("Starting thread {}", thread_number);
+            trace!("Starting thread {thread_number}");
             let cond_sync_t = cond_sync.clone();
             worker_handles.push(
                 std::thread::Builder::new()
@@ -87,7 +87,7 @@ mod d {
     }
 
     fn do_work(thread_number: usize, cond_sync: CondSync<usize>) {
-        trace!("({})     Thread started working", thread_number);
+        trace!("({thread_number})     Thread started working");
         trace!("ERROR_IF_PRINTED");
 
         cond_sync
@@ -95,7 +95,7 @@ mod d {
             .unwrap();
 
         for idx in 0..NO_OF_LOGLINES_PER_THREAD {
-            debug!("({})  writing out line number {}", thread_number, idx);
+            debug!("({thread_number})  writing out line number {idx}");
         }
         trace!("MUST_BE_PRINTED");
     }