+++ tests/test_multi_threaded_dates.rs
@@ -57,9 +57,9 @@ fn test_multi_threaded_dates() {
 // Starts given number of worker threads and lets each execute `do_work`
 fn start_worker_threads(no_of_workers: usize, cond_sync: &CondSync<usize>) -> Vec<JoinHandle<u8>> {
     let mut worker_handles: Vec<JoinHandle<u8>> = Vec::with_capacity(no_of_workers);
-    trace!("Starting {} worker threads", no_of_workers);
+    trace!("Starting {no_of_workers} worker threads");
     for thread_number in 0..no_of_workers {
-        trace!("Starting thread {}", thread_number);
+        trace!("Starting thread {thread_number}");
         let cond_sync_t = cond_sync.clone();
         worker_handles.push(
             std::thread::Builder::new()
@@ -76,7 +76,7 @@ fn start_worker_threads(no_of_workers: usize, cond_sync: &CondSync<usize>) -> Ve
 }
 
 fn do_work(thread_number: usize, cond_sync: CondSync<usize>) {
-    trace!("({})     Thread started working", thread_number);
+    trace!("({thread_number})     Thread started working");
     trace!("ERROR_IF_PRINTED");
 
     cond_sync
@@ -84,7 +84,7 @@ fn do_work(thread_number: usize, cond_sync: CondSync<usize>) {
         .unwrap();
 
     for idx in 0..NO_OF_LOGLINES_PER_THREAD {
-        debug!("({})  writing out line number {}", thread_number, idx);
+        debug!("({thread_number})  writing out line number {idx}");
     }
     trace!("MUST_BE_PRINTED");
 }